logic broken out into files by duty
