+++ example_test.go
@@ -1,16 +1,15 @@
 // Copyright...
 
 // This example demonstrates opening a Connection and doing some basic operations.
-package swift_test
+package swift
 
 import (
 	"fmt"
-	"github.com/ncw/swift"
 )
 
 func Example() {
 	// Create a connection
-	c := swift.Connection{
+	c := Connection{
 		UserName: "user",
 		ApiKey:   "key",
 		AuthUrl:  "auth_url",
@@ -30,7 +29,7 @@ var container string
 
 func ExampleConnection_ObjectsWalk() {
 	objects := make([]string, 0)
-	err := c.ObjectsWalk(container, nil, func(opts *swift.ObjectsOpts) (interface{}, error) {
+	err := c.ObjectsWalk(container, nil, func(opts *ObjectsOpts) (interface{}, error) {
 		newObjects, err := c.ObjectNames(container, opts)
 		if err == nil {
 			objects = append(objects, newObjects...)
